完成数据库0.1.0.alpha版本和配套异常库的编写

+++ SpecialCoolManager/SCM_Database.py
@@ -1,6 +1,6 @@
 """
-内部版本号：0.1.0.pro-alpha
-目前状态：工作中
+内部版本号：0.1.0.alpha
+目前状态：发布测试版，进行功能更新和bug修复
 
 模块说明：负责软件数据库相关操作
 """
@@ -11,13 +11,13 @@ import SCM_Exception
 
 
 class SQLDBMethod:
-    """
-    类属性
-    database_estorage 数据库储存文件夹地址
-    db_name 数据库名称，不带.db后缀
-
-    类方法
-    get_con() 返回数据库的connect类
+    """  # TODO(急迫) 统一文档字符串格式
+    **类属性** \n
+    *database_estorage -> str* 数据库储存文件夹地址，默认为相对文件夹 /Database \n
+    *db_name -> str* 数据库文件名称，不带.db后缀 \n
+    \n
+    **类方法** \n
+    *get_con()* 返回指定数据库的connect类 \n
     """
 
     def __init__(self):
@@ -28,26 +28,29 @@ class SQLDBMethod:
         if not os.path.exists(self.database_estorage):
             os.mkdir(self.database_estorage)
         if self.db_name == '':
-            connect_back = None  # 此分支会引发方法不存在的异常，注意处理
+            connect_back = None  # 此条件分支会引发AttributeError异常，请注意捕获
         else:
             connect_back = sqlite3.connect(f'{self.database_estorage}/{self.db_name}.db')
         return connect_back
 
 
-class SQLTableMethod:  # TODO(长期) 此类可能会触发方法不存在的异常
+class SQLTableMethod:  # TODO(长期) 此类可能会触发AttributeError异常
     """
-    类属性
-    table_name 指定的表名
-
-    类方法
-    con_close() 数据库con的提交和断开连接功能
-    table_check() 检查表名是否存在
-    table_create(table_format) 创建指定table_format格式的表
-    table_create() 删除指定的表
-
-    类异常
-    SCM_Exception.TableExist
-    SCM_Exception.TableInexist
+    类的初始化\n
+    get_con -> sqlite.connect 传入指定数据库的connect类\n
+    \n
+    类属性\n
+    table_name -> str 指定的表名\n
+    \n
+    类方法\n
+    con_close() 数据库con的提交和断开连接功能，注意最后必须调用\n
+    table_check() 检查指定的表名是否存在\n
+    table_create(table_format -> str) 创建指定table_format格式的表\n
+    table_create() 删除指定的表\n
+    \n
+    类异常\n
+    SCM_Exception.TableExist\n
+    SCM_Exception.TableInexist\n
     """
 
     def __init__(self, get_con):
@@ -84,13 +87,31 @@ class SQLTableMethod:  # TODO(长期) 此类可能会触发方法不存在的异
             self._cur.execute(f"DROP TABLE {self.table_name}")
         return
 
-    def _table_backup(self):  # TODO(长期) 完善数据库备份功能
+    def _table_backup(self):  # TODO(长期) 添加数据库备份功能
         pass
 
 
 class SQLColumnMethod:
     """
-    pass
+    类的初始化\n
+    get_con -> sqlite.connect 传入指定数据库的connect类\n
+    \n
+    类属性\n
+    table_name -> str 指定的表名\n
+    column_name -> tup 指定的列名，传入元组\n
+    sql_where -> str 不带WHERE的SQL的where语句\n
+    \n
+    类方法\n
+    column_insert(column_value -> tup) 在指定列插入传入的column_value数据\n
+    column_update(column_value -> tup) 更新指定列的数据为传入的column_value，WHERE语句强制必须添加\n
+    column_delete() 删除WHERE符合条件的表，WHERE语句强制必须添加\n
+    column_value_get_one(use_where=False) 返回单个搜索结果，use_where用于控制WHERE语句的启用，没有为空列表\n
+    column_value_get_many(size -> int, use_where=False) 返回size个搜索结果，use_where用于控制WHERE语句的启用\n
+    column_value_get_all(use_where=False) 返回所有的搜索结果，use_where用于控制WHERE语句的启用\n
+    \n
+    类异常\n
+    SCM_Exception.ColumnInsertMore
+    SCM_Exception.ColumnInsertLess
     """
 
     def __init__(self, get_con):
@@ -105,13 +126,13 @@ class SQLColumnMethod:
         table_check = SQLTableMethod(self._con)
         table_check.table_name = self.table_name
         table_check_bool = table_check.table_check()
-        if not table_check_bool:
+        if not table_check_bool:  # 表名必须存在
             raise SCM_Exception.ColumnTableError
         else:
             return True
 
     def _column_name_getin(self):
-        """将以列表格式传入的列名转换为SQL语句"""
+        """将以列表格式传入的列名转换为合法的SQL语句"""
         if not self.column_name:  # 缺省时默认为所有的列
             column_name_sqlin = '*'
         else:
@@ -124,6 +145,13 @@ class SQLColumnMethod:
 
     def column_insert(self, column_value):
         self._table_check()
+        column_name_len = len(self.column_name)
+        column_value_len = len(column_value)
+        if column_name_len != column_value_len:
+            if column_name_len < column_value_len:
+                raise SCM_Exception.ColumnInsertMore
+            elif column_name_len > column_value_len:
+                raise SCM_Exception.ColumnInsertLess
         value_placeholder = f"({('?, ' * len(column_value))[:-2]})"
         column_name = self._column_name_getin()
         if column_name == '*':
@@ -132,9 +160,9 @@ class SQLColumnMethod:
         return
 
     def _column_select(self, use_where=False):
-        """返回执行搜索的SQL语句"""
-        if use_where:
-            sql_where = f' WHERE {self.sql_where}'
+        """返回正在执行搜索的SQL搜索语句"""
+        if use_where and self.sql_where != '':
+            sql_where = f' WHERE {self.sql_where}'  # 保留WHERE前的空格
         else:
             sql_where = ''
         return self._cur.execute(f"SELECT {self._column_name_getin()[1: -1]} FROM {self.table_name} {sql_where}")
@@ -143,18 +171,40 @@ class SQLColumnMethod:
         column_name_in = ''
         for i in self.column_name:
             column_name_in += f"{i}=?, "
+        column_name_len = len(self.column_name)
+        column_value_len = len(column_value)
+        if column_name_len != column_value_len:
+            if column_name_len < column_value_len:
+                raise SCM_Exception.ColumnInsertMore
+            elif column_name_len > column_value_len:
+                raise SCM_Exception.ColumnInsertLess
+        if self.sql_where == '':
+            raise SCM_Exception.ColumnUpdateWhereNull
         self._cur.execute(f"UPDATE {self.table_name} SET {column_name_in[:-2]} WHERE {self.sql_where}", column_value)
         return
 
     def column_delete(self):
+        if self.sql_where == '':
+            raise SCM_Exception.ColumnDeleteNull
         self._cur.execute(f"DELETE FROM {self.table_name} WHERE {self.sql_where}")
         return
 
-    def column_value_get(self):
-        pass
+    def column_value_get_one(self, use_where=False):
+        value_get = self._column_select(use_where).fetchone()
+        if not value_get:
+            value_get = []
+        return value_get
+
+    def column_value_get_many(self, size, use_where=False):
+        value_get = self._column_select(use_where).fetchmany(size)
+        return value_get
+
+    def column_value_get_all(self, use_where=False):
+        value_get = self._column_select(use_where).fetchall()
+        return value_get
 
 
-class SQLWhereMake:  # TODO(临时) 完成where语句编写
+class SQLWhereMake:  # TODO(长期) 完成where语句编写
     pass
 
 
@@ -163,13 +213,12 @@ if __name__ == '__main__':
     test_1.database_estorage = r'D:\Programs\Programs\Working\Special_Cool_Manager\test'
     test_1.db_name = 'database'
     test_con = test_1.get_con()
+    print(type(test_con))
 
-    test_2 = SQLTableMethod(test_con)
-    # test_2.table_name = 'test'
-    # test_2.table_create('test_1 TEXT')
+    test_2 = SQLColumnMethod(test_con)
+    test_2.table_name = 'test_1'
+    # test_2.sql_where = "name_1 = 2"
+    test_2.column_name = ('name_2', 'name_1')
+    print(test_2.column_value_get_all(use_where=True))
 
-    test_3 = SQLColumnMethod(test_con)
-    test_3.table_name = 'test'
-    test_3.column_insert(('1234', ))
-
-    test_2.con_close()
+    SQLTableMethod(test_con).con_close()
